api/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pandas==2.2.2
openpyxl==3.1.2
supabase==2.3.0
google-generativeai==0.3.2
unidecode==1.3.7
python-dotenv==1.0.0